Support saving files/blocks on disk.
This first adds Save/Load functions to filestore/blockstore to save the metadata. Files and blocks/shards are stored in separate files, with references between them handled using FileId/BlockId.

Additionally, a test is added that saves a peer data, kills it, and recreates it to see if file can still be recovered. To support killing a peer, FluidBackup now supports Stop function.